refactor: move sync and hooks to store folder
