+++ apps/tlon-mobile/src/screens/ChatListScreen.tsx
@@ -1,6 +1,6 @@
 import type { NativeStackScreenProps } from '@react-navigation/native-stack';
 import type * as db from '@tloncorp/shared/dist/db';
-import * as hooks from '@tloncorp/shared/dist/hooks';
+import * as store from '@tloncorp/shared/dist/store';
 import { ChatList, ChatOptionsSheet, View } from '@tloncorp/ui';
 import React from 'react';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
@@ -15,7 +15,7 @@ type ChatListScreenProps = NativeStackScreenProps<
 export default function ChatListScreen(props: ChatListScreenProps) {
   const [longPressedItem, setLongPressedItem] =
     React.useState<db.Channel | null>(null);
-  const { data: chats } = hooks.useCurrentChats();
+  const { data: chats } = store.useCurrentChats();
   const insets = useSafeAreaInsets();
 
   return (
