+++ packages/shared/src/api/index.ts
@@ -9,5 +9,5 @@ export * from './unreadsApi';
 export * from './userApi';
 export * from './urbit';
 export * from './postsApi';
-export * from './reactQuery';
+export * from '../store/reactQuery';
 export * from './initApi';
