+++ packages/shared/src/index.ts
@@ -10,8 +10,8 @@ export type {
 } from './types/native';
 export { parseActiveTab, trimFullPath } from './logic/navigation';
 export * from './logic/utils';
-export * from './hooks';
-export * as sync from './sync';
+export * from './store';
+export * as sync from './store/sync';
 export * as utils from './logic/utils';
 export * as tiptap from './logic/tiptap';
 export * from './debug';
