+++ packages/shared/src/store/sync.test.ts
@@ -1,8 +1,17 @@
 import * as $ from 'drizzle-orm';
 import { beforeAll, beforeEach, expect, test, vi } from 'vitest';
 
-import { toClientGroup, toPagedPostsData } from './api';
-import * as db from './db';
+import { toClientGroup, toPagedPostsData } from '../api';
+import * as db from '../db';
+import rawChannelPostsData from '../test/channelPosts.json';
+import rawContactsData from '../test/contacts.json';
+import rawGroupsData from '../test/groups.json';
+import rawGroupsInitData from '../test/groupsInit.json';
+import { getClient, resetDb, setupDb } from '../test/helpers';
+import { setScryOutput, setScryOutputs } from '../test/helpers';
+import { GroupsInit, PagedPosts } from '../urbit';
+import { Contact as UrbitContact } from '../urbit/contact';
+import { Group as UrbitGroup } from '../urbit/groups';
 import {
   syncChannel,
   syncContacts,
@@ -11,15 +20,6 @@ import {
   syncInitData,
   syncPinnedItems,
 } from './sync';
-import rawChannelPostsData from './test/channelPosts.json';
-import rawContactsData from './test/contacts.json';
-import rawGroupsData from './test/groups.json';
-import rawGroupsInitData from './test/groupsInit.json';
-import { getClient, resetDb, setupDb } from './test/helpers';
-import { setScryOutput, setScryOutputs } from './test/helpers';
-import { GroupsInit, PagedPosts } from './urbit';
-import { Contact as UrbitContact } from './urbit/contact';
-import { Group as UrbitGroup } from './urbit/groups';
 
 const contactsData = rawContactsData as unknown as Record<string, UrbitContact>;
 const groupsData = rawGroupsData as unknown as Record<string, UrbitGroup>;
