+++ packages/shared/tsup.config.ts
@@ -7,10 +7,9 @@ export default defineConfig({
     'src/urbit/*',
     'src/client/index.ts',
     'src/db/index.ts',
-    'src/hooks/index.ts',
     'src/db/migrations/index.ts',
     'src/api/index.ts',
-    'src/sync.ts',
+    'src/store/index.ts',
   ],
   format: ['esm'],
   minify: false,
